context: add context to buildpack installations
The primary reason for this change is to ensure that `@shared/context` is installed and built as a part of the buildpack compile step.

In addition to this, because all packages are now part of `yarn workspaces`, the individual installations have been replaced with a single `yarn install`. Individual builds are still required, but there is no longer a need to run `yarn install` at the package level.